+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **http client factory:** add default options to reduce the amount of config required (when not specified).
 - **http client factory:** `Add` now returns newly created `FluentHttpClient`.
 - **http client factory:** implement `WithRequestBuilderDefaults` additional hook in order customize `FluentHttpRequestBuilder` on creation.
+- **http client factory:** implement `WithMessageHandler` in order to replace the HTTP stack for the HTTP client.
 
 - **http client builder:** add new method `AddMiddleware(Type)`.
 - **http client builder:** align naming with request builder.